refactor(keyword-detector): split constants into domain-specific modules
