+++ src/hooks/keyword-detector/analyze/index.ts
@@ -1,0 +1,1 @@
+export { ANALYZE_PATTERN, ANALYZE_MESSAGE } from "./default"
