+++ src/hooks/keyword-detector/detector.ts
@@ -17,26 +17,27 @@ export function removeCodeBlocks(text: string): string {
  * Resolves message to string, handling both static strings and dynamic functions.
  */
 function resolveMessage(
-  message: string | ((agentName?: string) => string),
-  agentName?: string
+  message: string | ((agentName?: string, modelID?: string) => string),
+  agentName?: string,
+  modelID?: string
 ): string {
-  return typeof message === "function" ? message(agentName) : message
+  return typeof message === "function" ? message(agentName, modelID) : message
 }
 
-export function detectKeywords(text: string, agentName?: string): string[] {
+export function detectKeywords(text: string, agentName?: string, modelID?: string): string[] {
   const textWithoutCode = removeCodeBlocks(text)
   return KEYWORD_DETECTORS.filter(({ pattern }) =>
     pattern.test(textWithoutCode)
-  ).map(({ message }) => resolveMessage(message, agentName))
+  ).map(({ message }) => resolveMessage(message, agentName, modelID))
 }
 
-export function detectKeywordsWithType(text: string, agentName?: string): DetectedKeyword[] {
+export function detectKeywordsWithType(text: string, agentName?: string, modelID?: string): DetectedKeyword[] {
   const textWithoutCode = removeCodeBlocks(text)
   const types: Array<"ultrawork" | "search" | "analyze"> = ["ultrawork", "search", "analyze"]
   return KEYWORD_DETECTORS.map(({ pattern, message }, index) => ({
     matches: pattern.test(textWithoutCode),
     type: types[index],
-    message: resolveMessage(message, agentName),
+    message: resolveMessage(message, agentName, modelID),
   }))
     .filter((result) => result.matches)
     .map(({ type, message }) => ({ type, message }))
