+++ src/hooks/keyword-detector/index.ts
@@ -35,7 +35,8 @@ export function createKeywordDetectorHook(ctx: PluginInput, collector?: ContextC
 
       // Remove system-reminder content to prevent automated system messages from triggering mode keywords
       const cleanText = removeSystemReminders(promptText)
-      let detectedKeywords = detectKeywordsWithType(removeCodeBlocks(cleanText), currentAgent)
+      const modelID = input.model?.modelID
+      let detectedKeywords = detectKeywordsWithType(removeCodeBlocks(cleanText), currentAgent, modelID)
 
       if (isPlannerAgent(currentAgent)) {
         detectedKeywords = detectedKeywords.filter((k) => k.type !== "ultrawork")
