+++ src/hooks/keyword-detector/search/index.ts
@@ -1,0 +1,1 @@
+export { SEARCH_PATTERN, SEARCH_MESSAGE } from "./default"
